move EpochOutputStore out of contrib keeping BC
