+++ ignite/handlers/__init__.py
@@ -4,6 +4,7 @@ from ignite.engine import Engine
 from ignite.engine.events import Events
 from ignite.handlers.checkpoint import Checkpoint, DiskSaver, ModelCheckpoint
 from ignite.handlers.early_stopping import EarlyStopping
+from ignite.handlers.stores import EpochOutputStore
 from ignite.handlers.terminate_on_nan import TerminateOnNan
 from ignite.handlers.time_limit import TimeLimit
 from ignite.handlers.timing import Timer
@@ -17,6 +18,7 @@ __all__ = [
     "TerminateOnNan",
     "global_step_from_engine",
     "TimeLimit",
+    "EpochOutputStore",
 ]
